+++ frontend/app/features/blogs/components/BlogPageMediaGallery.tsx
